More efficient GroupSeparationExperiment by caching representations.
Before, representations of the same model were computed multiple times for the same data and all kept in memory. Now they refer to the same object, if they are needed multiple times. This should massively save memory and make the previous manual deletion of them unnecessary as well as speed up the whole experiment.